% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{\code{fars_read} checks if the file exists and then load it as data.frame.
If the file does not exist function will generate an error.}
\source{
Coursera Mastering Software Development in R - Johns Hopkins University
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character - a path to the file or the file name}
}
\value{
load data in a data.frame format
}
\description{
\code{fars_read} checks if the file exists and then load it as data.frame.
If the file does not exist function will generate an error.
}
\examples{
\dontrun{
   fars_read(filename)
}

}
\seealso{
\code{\link{make_filename}} \code{\link{fars_read_years}} \code{\link{fars_summarize_years}} \code{\link{fars_map_state}}
}
