% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{\code{fars_read_years} takes a list of years and returns a list of data.frames with month and year columns based on provided files.
the files need to be in the working directory.}
\source{
Coursera Mastering Software Development in R - Johns Hopkins University
}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list or vector of integers/characters - years of origin}
}
\value{
a list of data.frames with month and year columns
}
\description{
\code{fars_read_years} takes a list of years and returns a list of data.frames with month and year columns based on provided files.
the files need to be in the working directory.
}
\examples{
\dontrun{
   fars_read_years(year:(year+n))
   fars_read_years(list(year,(year+n)))
   fars_read_years(year)
   fars_read_years(list(year:(year+n))) #example of error (return NULL)
}

}
\seealso{
\code{\link{make_filename}} \code{\link{fars_read}} \code{\link{fars_summarize_years}} \code{\link{fars_map_state}}
}
