% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{\code{fars_summarize_years} return the data.frame with total number of accidents for each month.
the files need to be in the working directory.}
\source{
Coursera Mastering Software Development in R - Johns Hopkins University
}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list or vector of integers/characters - years of origin}
}
\value{
a data.frame in a wide format with total number of accidents for each month
}
\description{
\code{fars_summarize_years} return the data.frame with total number of accidents for each month.
the files need to be in the working directory.
}
\examples{
\dontrun{
   fars_summarize_years(2013:2015)
   fars_summarize_years(list(2014:2015)) #example of error
}

}
\seealso{
\code{\link{make_filename}} \code{\link{fars_read_years}} \code{\link{fars_read}} \code{\link{fars_map_state}}
}
