% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{list of codes for US states}
\format{A data frame with 53 rows and 2 variables:
\describe{
  \item{nr}{code for a state}
  \item{name}{name of a state}
}}
\source{
\url{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/812315}
}
\usage{
us_states
}
\description{
STATE variable from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
which is a nationwide census providing the American public yearly data regarding fatal injuries suffered
in motor vehicle traffic crashes.
}
\keyword{datasets}
